requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-genai==0.3.0
pydantic==2.6.4
httpx==0.27.0